Move Game class to a separate file
